Создан компонент для списка городов Центр карты перерисовывается при выборе нового города
